Add a clinic

+++ src/main/java/org/launchcode/VetConnect/controllers/AuthenticationController.java
@@ -26,7 +26,7 @@ public class AuthenticationController {
     private static final String userSessionKey = "user";
 
     public User getUserFromSession(HttpSession session) {
-        Integer userId = (Integer) session.getAttribute(userSessionKey);
+        Long userId = (Long) session.getAttribute(userSessionKey);
         if (userId == null) {
             return null;
         }
